examples/addons_examples/...: Now the demos can be compiled with IMGUI_DISABLE_OBSOLETE_FUNCTIONS defined. examples/addons_examples/main.cpp: Removed the 'Splitter test' button (it did not work anymore and it was not very useful). addons/imguinodegrapheditor/imguinodegrapheditor.cpp: This is probably the biggest change, to make it more compatible with the new Dear ImGui data types (e.g. DragScalar(...), SliderScalar(...)). Expect issues here... addons/imguinodegrapheditor/imguinodegrapheditor.cpp: Fixed (in a bad way) mouse rectangular selection of nodes that was, for some strange reason, broken.

+++ addons/imguibindings/imguibindings.cpp
@@ -643,11 +643,11 @@ ImTextureID ImImpl_LoadTextureFromMemory(const unsigned char* filenameInMemory,i
 
 ImTextureID ImImpl_LoadTexture(const char* filename, int req_comp, bool useMipmapsIfPossible, bool wraps, bool wrapt,bool minFilterNearest,bool magFilterNearest)  {
     // We avoid using stbi_load(...), because we support UTF8 paths under Windows too.
-    int file_size = 0;
+    size_t file_size = 0;
     unsigned char* file = (unsigned char*) ImFileLoadToMemory(filename,"rb",&file_size,0);
     ImTextureID texId = NULL;
     if (file)   {
-        texId = ImImpl_LoadTextureFromMemory(file,file_size,req_comp,useMipmapsIfPossible,wraps,wrapt,minFilterNearest,magFilterNearest);
+        texId = ImImpl_LoadTextureFromMemory(file,(int)file_size,req_comp,useMipmapsIfPossible,wraps,wrapt,minFilterNearest,magFilterNearest);
         ImGui::MemFree(file);file=NULL;
     }
     return texId;
